Load various gene sets

Load various gene sets

Usage

load_go_genesets(org_db, ontology = "BP")

load_kegg_genesets(organism, db = "pathway")

load_msigdb_genesets(collection = "h.all", version = "2024.1.Hs")

load_reactome_genesets(organism)

load_simona_genesets(fun, ontology, organism = "human")

Arguments

org_db

An OrgDb object.

ontology

Just a tag, please use "hp" or "rdo".

organism

See Details.

db

A KEGG database. The value can be one of "pathway", "module", "ko", "network", "disease" and "drug".

collection

Collection of the MSigDB gene sets. All possible values can be found via list_msigdb_versions().

version

Version of the MSigDB database. All possible values can be found via list_msigdb_collections().

fun

Use simona::ontology_hp or simona::ontology_rdo.

Details

Genes are all in the EntreZ ID type. All the gene sets are saved in cache.

The value should be set differently for specific fgsea_*() functions.

for load_kegg_genesets(), the value should be a KEGG organism code, such as "hsa" or "mmu".

for load_reactome_genesets(), the value should a prefix of the Reactome pathway ID that represents the organism. E.g. "HSA" for human.

for load_simona_genesets(), the value can only be one of "human", "mouse" and "rat".

All valid values for load_reactome_genesets() are:

c("BTA", "CEL", "CFA", "DRE", "DDI", "DME", "GGA", "HSA", "MMU",
  "MTU", "PFA", "RNO", "SCE", "SPO", "SSC", "XTR")

The human phenotype gene sets and disease gene sets are supported by the simona package.