Over-representation analysis

Over-representation analysis

Usage

ora(genes, gs, universe = NULL, min_hits = 3, min_size = 5, max_size = 2500)

ora_go(genes, org_db = org.Hs.eg.db::org.Hs.eg.db, ontology = "BP", ...)

ora_kegg(genes, organism = "hsa", db = "pathway", ...)

ora_msigdb(genes, collection = "h.all", version = "2024.1.Hs", ...)

ora_reactome(genes, organism = "HSA", ...)

ora_keywords(genes, organism = "human", ...)

ora_phenotype(genes, organism = "human", ...)

ora_disease(genes, organism = "human", ...)

Arguments

genes

A vector of genes.

gs

A list of gene sets. Genes should have the smae ID type as in genes.

universe

A vector of background genes.

min_hits

Minimal number of overlapping genes in genes and gene sets.

min_size

Minimal number of genes in gene sets.

max_size

Maximal number of genes in gene sets.

org_db

An OrgDb object for the organism. It can be from org.*.db packages or downloaded by the AnnotationHub package.

ontology

Namespace of GO. Value should be one of "BP", "CC" or "MF".

...

Passed to ora().

organism

See Details.

db

A KEGG database. The value can be one of "pathway", "module", "ko", "network", "disease" and "drug".

collection

Collection of the MSigDB gene sets. All possible values can be found via list_msigdb_versions().

version

Version of the MSigDB database. All possible values can be found via list_msigdb_collections().

Details

The value should be set differently for specific ora_*() functions.

for ora_kegg(), the value should be a KEGG organism code, such as "hsa" or "mmu".

for ora_reactome(), the value should a prefix of the Reactome pathway ID that represents the organism. E.g. "HSA" for human.

for ora_keywords(), the value can be a organism name, e.g. "human", the latin name or the taxon ID.

for ora_phenotype() and fgsea_disease(), the value can only be one of "human", "mouse" and "rat".